Working on implementing the working method into this one
