+++ client/hooks/useSignup.js
@@ -10,7 +10,7 @@ export const useSignup = () => {
     setIsLoading(true)
     setError(null)
 
-    const response = await fetch('/user/signup', {
+    const response = await fetch('http://localhost:8080/api/users/signup', {
       method: 'POST',
       headers: {'Content-Type': 'application/json'},
       body: JSON.stringify({ name, email, password })
