--- server/controllers/userController.js
@@ -1,41 +1,0 @@
-//const User = require('../models/userModel.js');
-import * as User from './../models/userModel.js';
-const jwt = require('jsonwebtoken');
-
-const createToken = (_id) => {
-  return jwt.sign({ _id }, process.env.SECRET, { expiresIn: '3d' });
-};
-
-// login a user
-const loginUser = async (req, res) => {
-  const { email, password } = req.body;
-
-  try {
-    const user = await User.login(email, password);
-
-    // create a token
-    const token = createToken(user._id);
-
-    res.status(200).json({ email, token });
-  } catch (error) {
-    res.status(400).json({ error: error.message });
-  }
-};
-
-// signup a user
-const signupUser = async (req, res) => {
-  const { name, email, password } = req.body;
-
-  try {
-    const user = await User.signup(name, email, password);
-
-    // create a token
-    const token = createToken(user._id);
-
-    res.status(200).json({ email, token });
-  } catch (error) {
-    res.status(400).json({ error: error.message });
-  }
-};
-
-module.exports = { loginUser, signupUser };
