+++ server/models/userModel.js
@@ -70,4 +70,5 @@ userSchema.statics.login = async function(email, password) {
   return user;
 };
 
-module.exports = mongoose.model('User', userSchema);
+//module.exports = mongoose.model('User', userSchema);
+export default userSchema;
