+++ server/routes/userRoutes.js
@@ -1,16 +1,57 @@
 import express from 'express';
+import * as User from './../models/userModel.js';
+import jwt from 'jsonwebtoken';
 
-import * as user from './../controllers/userController.js';
-
-// controller functions
-//const { loginUser, signupUser } = require('../controllers/userController.js');
+const createToken = (_id) => {
+  return jwt.sign({ _id }, process.env.SECRET, { expiresIn: '3d' });
+};
 
 const router = express.Router();
 
 // login route
-router.post('/login', user.loginUser);
+router.post('/login', async (req, res) => {
+  const { email, password } = req.body;
+
+  try {
+    const user = await User.login(email, password);
+
+    // create a token
+    const token = createToken(user._id);
+
+    res.status(200).json({ email, token });
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }});
 
 // signup route
-router.post('/signup', user.signupUser);
+router.post('/signup', async (req, res) => {
+  const { name, email, password } = req.body;
 
-module.exports = { router };
+  try {
+    const user = await User.signup(name, email, password);
+
+    // create a token
+    const token = createToken(user._id);
+
+    res.status(200).json({ email, token });
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }});
+
+/*router.post('/register', async (req, res) => {
+  const { name, email, password } = req.body;
+
+  try {
+    // Create a new user
+    const user = new User({ name, email, password });
+    await user.save();
+
+    res.json({ success: true, message: 'User registered successfully' });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ success: false, message: 'An error occurred' });
+  }
+});*/
+
+export default router;
+
